+++ yandex/data_source_yandex_api_gateway.go
@@ -41,6 +41,13 @@ func dataSourceYandexApiGateway() *schema.Resource {
 				Set:      schema.HashString,
 			},
 
+			"user_domains": {
+				Type:     schema.TypeSet,
+				Computed: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Set:      schema.HashString,
+			},
+
 			"domain": {
 				Type:     schema.TypeString,
 				Computed: true,