+++ yandex/resource_yandex_api_gateway.go
@@ -2,11 +2,9 @@ package yandex
 
 import (
 	"fmt"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/hashcode"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/yandex-cloud/go-genproto/yandex/cloud/serverless/apigateway/v1"
 	"google.golang.org/genproto/protobuf/field_mask"
-	"io/ioutil"
 	"strings"
 	"time"
 )
@@ -58,9 +56,11 @@ func resourceYandexApiGateway() *schema.Resource {
 				Required: true,
 			},
 
-			"spec_content_hash": {
-				Type:     schema.TypeInt,
-				Optional: true,
+			"user_domains": {
+				Type:     schema.TypeSet,
+				Computed: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Set:      schema.HashString,
 			},
 
 			"domain": {
@@ -92,7 +92,7 @@ func resourceYandexApiGatewayCreate(d *schema.ResourceData, meta interface{}) er
 	ctx, cancel := config.ContextWithTimeout(d.Timeout(schema.TimeoutCreate))
 	defer cancel()
 
-	req, specContent, err := getCreateApiGatewayRequest(d, config)
+	req, err := getCreateApiGatewayRequest(d, config)
 	if err != nil {
 		return err
 	}
@@ -113,32 +113,25 @@ func resourceYandexApiGatewayCreate(d *schema.ResourceData, meta interface{}) er
 	}
 
 	d.SetId(md.ApiGatewayId)
+	d.Set("spec", d.Get("spec").(string))
 
 	err = op.Wait(ctx)
 	if err != nil {
 		return fmt.Errorf("Error while requesting API to create Yandex Cloud API Gateway: %s", err)
 	}
 
-	specHash := hashcode.String(*specContent)
-	setSpecHash(d, specHash)
-
 	return resourceYandexApiGatewayRead(d, meta)
 }
 
-func getCreateApiGatewayRequest(d *schema.ResourceData, config *Config) (*apigateway.CreateApiGatewayRequest, *string, error) {
+func getCreateApiGatewayRequest(d *schema.ResourceData, config *Config) (*apigateway.CreateApiGatewayRequest, error) {
 	labels, err := expandLabels(d.Get("labels"))
 	if err != nil {
-		return nil, nil, fmt.Errorf("Error expanding labels while creating Yandex Cloud API Gateway: %s", err)
+		return nil, fmt.Errorf("Error expanding labels while creating Yandex Cloud API Gateway: %s", err)
 	}
 
 	folderID, err := getFolderID(d, config)
 	if err != nil {
-		return nil, nil, fmt.Errorf("Error getting folder ID while creating Yandex Cloud API Gateway: %s", err)
-	}
-
-	specContent, err := loadSpec(d.Get("spec").(string))
-	if err != nil {
-		return nil, nil, fmt.Errorf("Error getting spec while creating Yandex Cloud API Gateway: %s", err)
+		return nil, fmt.Errorf("Error getting folder ID while creating Yandex Cloud API Gateway: %s", err)
 	}
 
 	req := &apigateway.CreateApiGatewayRequest{
@@ -147,16 +140,17 @@ func getCreateApiGatewayRequest(d *schema.ResourceData, config *Config) (*apigat
 		Description: d.Get("description").(string),
 		Labels:      labels,
 		Spec: &apigateway.CreateApiGatewayRequest_OpenapiSpec{
-			OpenapiSpec: *specContent,
+			OpenapiSpec: d.Get("spec").(string),
 		},
 	}
-	return req, specContent, err
+
+	return req, err
 }
 
 func resourceYandexApiGatewayUpdate(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
 
-	ctx, cancel := config.ContextWithTimeout(d.Timeout(schema.TimeoutCreate))
+	ctx, cancel := config.ContextWithTimeout(d.Timeout(schema.TimeoutUpdate))
 	defer cancel()
 
 	labels, err := expandLabels(d.Get("labels"))
@@ -167,6 +161,11 @@ func resourceYandexApiGatewayUpdate(d *schema.ResourceData, meta interface{}) er
 	d.Partial(true)
 
 	var updatePaths []string
+
+	if d.HasChange("spec") {
+		updatePaths = append(updatePaths, "spec")
+	}
+
 	if d.HasChange("name") {
 		updatePaths = append(updatePaths, "name")
 	}
@@ -179,17 +178,6 @@ func resourceYandexApiGatewayUpdate(d *schema.ResourceData, meta interface{}) er
 		updatePaths = append(updatePaths, "labels")
 	}
 
-	specContent, err := loadSpec(d.Get("spec").(string))
-	if err != nil {
-		return err
-	}
-	newSpecHash := hashcode.String(*specContent)
-	isSpecChanged := false
-	if d.Get("spec_content_hash") != newSpecHash {
-		updatePaths = append(updatePaths, "openapi_spec")
-		isSpecChanged = true
-	}
-
 	if len(updatePaths) != 0 {
 		req := apigateway.UpdateApiGatewayRequest{
 			ApiGatewayId: d.Id(),
@@ -198,7 +186,7 @@ func resourceYandexApiGatewayUpdate(d *schema.ResourceData, meta interface{}) er
 			Labels:       labels,
 			UpdateMask:   &field_mask.FieldMask{Paths: updatePaths},
 			Spec: &apigateway.UpdateApiGatewayRequest_OpenapiSpec{
-				OpenapiSpec: *specContent,
+				OpenapiSpec: d.Get("spec").(string),
 			},
 		}
 
@@ -209,23 +197,11 @@ func resourceYandexApiGatewayUpdate(d *schema.ResourceData, meta interface{}) er
 		}
 
 		for _, v := range updatePaths {
-			if v == "openapi_spec" {
-				continue
-			}
-
 			d.SetPartial(v)
-		}
-
-		if isSpecChanged {
-			d.SetPartial("spec_content_hash")
 		}
 	}
 
 	d.Partial(false)
-
-	if isSpecChanged {
-		setSpecHash(d, newSpecHash)
-	}
 
 	return resourceYandexApiGatewayRead(d, meta)
 }
@@ -267,15 +243,6 @@ func resourceYandexApiGatewayDelete(d *schema.ResourceData, meta interface{}) er
 	return nil
 }
 
-func loadSpec(filename string) (*string, error) {
-	fileBytes, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-	specString := string(fileBytes)
-	return &specString, nil
-}
-
 func flattenYandexApiGateway(d *schema.ResourceData, apiGateway *apigateway.ApiGateway) error {
 	createdAt, err := getTimestamp(apiGateway.CreatedAt)
 	if err != nil {
@@ -289,9 +256,11 @@ func flattenYandexApiGateway(d *schema.ResourceData, apiGateway *apigateway.ApiG
 	d.Set("domain", apiGateway.Domain)
 	d.Set("status", strings.ToLower(apiGateway.Status.String()))
 	d.Set("log_group_id", apiGateway.LogGroupId)
-	return d.Set("labels", apiGateway.Labels)
-}
 
-func setSpecHash(d *schema.ResourceData, hash int) {
-	d.Set("spec_content_hash", hash)
+	domains := make([]string, len(apiGateway.AttachedDomains))
+	for i, domain := range apiGateway.AttachedDomains {
+		domains[i] = domain.DomainId
+	}
+	d.Set("user_domains", convertStringArrToInterface(domains))
+	return d.Set("labels", apiGateway.Labels)
 }